비동기 씬 전환 테스트, 몬스터 데미지 적용

+++ Assets/Scripts/Interface/IDamageable.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void Hit(int damage);
+    void Hit(float damage);
 
     IEnumerator HitDelay();
 }
